+++ node/service/samples/registry_sample.js
@@ -4,39 +4,46 @@
 'use strict';
 
 var iothub = require('azure-iothub');
+var argv = require('yargs')
+             .usage('Usage: node $0 --connectionString <IOTHUB CONNECTION STRING>')
+             .demand(['connectionString'])
+             .alias('c', 'connectionString')
+             .describe('connectionString', 'IoT Hub service connection string.')
+             .argv
 
-var connectionString = '[IoT Connection String]';
-
-var registry = iothub.Registry.fromConnectionString(connectionString);
+var registry = iothub.Registry.fromConnectionString(argv.connectionString);
 
 // List devices
 console.log('**listing devices...');
 registry.list(function (err, deviceList) {
   deviceList.forEach(function (device) {
-    var key = device.authentication ? device.authentication.SymmetricKey.primaryKey : '<no primary key>';
-    console.log(device.deviceId + ': ' + key);
+    console.log(device.deviceId);
   });
 
   // Create a new device
   var device = new iothub.Device(null);
   device.deviceId = 'sample-device-' + Date.now();
   console.log('\n**creating device \'' + device.deviceId + '\'');
-  registry.create(device, printAndContinue('create', function next() {
-
+  registry.create(device, printAndContinue('create', function() {
     // Get the newly-created device
     console.log('\n**getting device \'' + device.deviceId + '\'');
-    registry.get(device.deviceId, printAndContinue('get', function next() {
-
+    registry.get(device.deviceId, printAndContinue('get', function() {
       // Delete the new device
       console.log('\n**deleting device \'' + device.deviceId + '\'');
-      registry.delete(device.deviceId, printAndContinue('delete'));
+      registry.delete(device.deviceId, printAndContinue('delete', function() {
+        process.exit(0);
+      }));
     }));
   }));
 });
 
 function printAndContinue(op, next) {
-  return function printResult(err, deviceInfo, res) {
-    if (err) console.log(op + ' error: ' + err.toString());
+  return function (err, deviceInfo, res) {
+    if (err) {
+      console.log(op + ' error: ' + err.toString());
+      process.exit(1);
+    }
+
     if (res) console.log(op + ' status: ' + res.statusCode + ' ' + res.statusMessage);
     if (deviceInfo) console.log(op + ' device info: ' + JSON.stringify(deviceInfo));
     if (next) next();